refactoring multiproc to fix deadlock
