--- nipype/pipeline/plugins/API.rst
@@ -1,8 +1,0 @@
-Execution plugin API
-====================
-
-Current status:
-
-class plugin_runner(PluginBase):
-
-    def run(graph, config, updatehash)
