+++ nipype/pipeline/plugins/tests/test_multiproc.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
-import logging
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+"""
+Test the resource management of MultiProc
+"""
 import os
-from multiprocessing import cpu_count
-
-import nipype.interfaces.base as nib
-from nipype.utils import draw_gantt_chart
-import nipype.pipeline.engine as pe
-from nipype.utils.profiler import log_nodes_cb, get_system_total_memory_gb
+import pytest
+from nipype.pipeline import engine as pe
+from nipype.interfaces import base as nib
 
 
 class InputSpec(nib.TraitedSpec):
@@ -33,11 +34,11 @@ class MultiprocTestInterface(nib.BaseInterface):
 
 
 def test_run_multiproc(tmpdir):
-    os.chdir(str(tmpdir))
+    tmpdir.chdir()
 
     pipe = pe.Workflow(name='pipe')
-    mod1 = pe.Node(interface=MultiprocTestInterface(), name='mod1')
-    mod2 = pe.MapNode(interface=MultiprocTestInterface(),
+    mod1 = pe.Node(MultiprocTestInterface(), name='mod1')
+    mod2 = pe.MapNode(MultiprocTestInterface(),
                       iterfield=['input1'],
                       name='mod2')
     pipe.connect([(mod1, mod2, [('output1', 'input1')])])
@@ -45,7 +46,7 @@ def test_run_multiproc(tmpdir):
     mod1.inputs.input1 = 1
     pipe.config['execution']['poll_sleep_duration'] = 2
     execgraph = pipe.run(plugin="MultiProc")
-    names = ['.'.join((node._hierarchy, node.name)) for node in execgraph.nodes()]
+    names = [node.fullname for node in execgraph.nodes()]
     node = list(execgraph.nodes())[names.index('pipe.mod1')]
     result = node.get_output('output1')
     assert result == [1, 1]
@@ -74,69 +75,13 @@ class SingleNodeTestInterface(nib.BaseInterface):
         return outputs
 
 
-def find_metrics(nodes, last_node):
-    """
-    """
-
-    # Import packages
-    from dateutil.parser import parse
-    import datetime
-
-    start = parse(nodes[0]['start'])
-    total_duration = max(int((parse(last_node['finish']) - start).total_seconds()), 1)
-
-    total_memory = []
-    total_threads = []
-    for i in range(total_duration):
-        total_memory.append(0)
-        total_threads.append(0)
-
-    now = start
-    for i in range(total_duration):
-        start_index = 0
-        node_start = None
-        node_finish = None
-
-        x = now
-
-        for j in range(start_index, len(nodes)):
-            node_start = parse(nodes[j]['start'])
-            node_finish = parse(nodes[j]['finish'])
-
-            if node_start < x and node_finish > x:
-                total_memory[i] += float(nodes[j]['estimated_memory_gb'])
-                total_threads[i] += int(nodes[j]['num_threads'])
-                start_index = j
-
-            if node_start > x:
-                break
-
-        now += datetime.timedelta(seconds=1)
-
-    return total_memory, total_threads
-
-
 def test_no_more_memory_than_specified(tmpdir):
     tmpdir.chdir()
-    LOG_FILENAME = tmpdir.join('callback.log').strpath
-    my_logger = logging.getLogger('callback')
-    my_logger.setLevel(logging.DEBUG)
-
-    # Add the log message handler to the logger
-    handler = logging.FileHandler(LOG_FILENAME)
-    my_logger.addHandler(handler)
-
-    max_memory = 1
     pipe = pe.Workflow(name='pipe')
-    n1 = pe.Node(interface=SingleNodeTestInterface(), name='n1')
-    n2 = pe.Node(interface=SingleNodeTestInterface(), name='n2')
-    n3 = pe.Node(interface=SingleNodeTestInterface(), name='n3')
-    n4 = pe.Node(interface=SingleNodeTestInterface(), name='n4')
-
-    n1.interface.estimated_memory_gb = 1
-    n2.interface.estimated_memory_gb = 1
-    n3.interface.estimated_memory_gb = 1
-    n4.interface.estimated_memory_gb = 1
+    n1 = pe.Node(SingleNodeTestInterface(), name='n1', mem_gb=1)
+    n2 = pe.Node(SingleNodeTestInterface(), name='n2', mem_gb=1)
+    n3 = pe.Node(SingleNodeTestInterface(), name='n3', mem_gb=1)
+    n4 = pe.Node(SingleNodeTestInterface(), name='n4', mem_gb=1)
 
     pipe.connect(n1, 'output1', n2, 'input1')
     pipe.connect(n1, 'output1', n3, 'input1')
@@ -144,87 +89,29 @@ def test_no_more_memory_than_specified(tmpdir):
     pipe.connect(n3, 'output1', n4, 'input2')
     n1.inputs.input1 = 1
 
-    pipe.run(plugin='MultiProc',
-             plugin_args={'memory_gb': max_memory,
-                          'status_callback': log_nodes_cb})
-
-    nodes = draw_gantt_chart.log_to_dict(LOG_FILENAME)
-    last_node = nodes[-1]
-    # usage in every second
-    memory, threads = find_metrics(nodes, last_node)
-
-    result = True
-    for m in memory:
-        if m > max_memory:
-            result = False
-            break
-
-    assert result
-
-    max_threads = cpu_count()
-
-    result = True
-    for t in threads:
-        if t > max_threads:
-            result = False
-            break
-
-    assert result,\
-        "using more threads than system has (threads is not specified by user)"
-
-    os.remove(LOG_FILENAME)
+    max_memory = 0.5
+    with pytest.raises(RuntimeError):
+        pipe.run(plugin='MultiProc',
+                 plugin_args={'memory_gb': max_memory,
+                              'n_procs': 2})
 
 
-def test_no_more_threads_than_specified():
-    LOG_FILENAME = 'callback.log'
-    my_logger = logging.getLogger('callback')
-    my_logger.setLevel(logging.DEBUG)
+def test_no_more_threads_than_specified(tmpdir):
+    tmpdir.chdir()
 
-    # Add the log message handler to the logger
-    handler = logging.FileHandler(LOG_FILENAME)
-    my_logger.addHandler(handler)
-
-    max_threads = 4
     pipe = pe.Workflow(name='pipe')
-    n1 = pe.Node(interface=SingleNodeTestInterface(), name='n1')
-    n2 = pe.Node(interface=SingleNodeTestInterface(), name='n2')
-    n3 = pe.Node(interface=SingleNodeTestInterface(), name='n3')
-    n4 = pe.Node(interface=SingleNodeTestInterface(), name='n4')
-
-    n1.interface.num_threads = 1
-    n2.interface.num_threads = 1
-    n3.interface.num_threads = 4
-    n4.interface.num_threads = 1
+    n1 = pe.Node(SingleNodeTestInterface(), name='n1', n_procs=2)
+    n2 = pe.Node(SingleNodeTestInterface(), name='n2', n_procs=2)
+    n3 = pe.Node(SingleNodeTestInterface(), name='n3', n_procs=4)
+    n4 = pe.Node(SingleNodeTestInterface(), name='n4', n_procs=2)
 
     pipe.connect(n1, 'output1', n2, 'input1')
     pipe.connect(n1, 'output1', n3, 'input1')
     pipe.connect(n2, 'output1', n4, 'input1')
     pipe.connect(n3, 'output1', n4, 'input2')
     n1.inputs.input1 = 4
-    pipe.config['execution']['poll_sleep_duration'] = 1
-    pipe.run(plugin='MultiProc', plugin_args={'n_procs': max_threads,
-                                              'status_callback': log_nodes_cb})
 
-    nodes = draw_gantt_chart.log_to_dict(LOG_FILENAME)
-    last_node = nodes[-1]
-    # usage in every second
-    memory, threads = find_metrics(nodes, last_node)
-
-    result = True
-    for t in threads:
-        if t > max_threads:
-            result = False
-            break
-
-    assert result, "using more threads than specified"
-
-    max_memory = get_system_total_memory_gb()
-    result = True
-    for m in memory:
-        if m > max_memory:
-            result = False
-            break
-    assert result,\
-        "using more memory than system has (memory is not specified by user)"
-
-    os.remove(LOG_FILENAME)
+    max_threads = 2
+    with pytest.raises(RuntimeError):
+        pipe.run(plugin='MultiProc',
+                 plugin_args={'n_procs': max_threads})
